// src/app/models/Employee.ts
export class Employee {
    EmployeeId: number = 0;
    FirstName: string = '';
    LastName: string = '';
    Email: string = '';
    Skills: string = '';
    PhoneNumber: string = '';
    Department: string = '';
    Position: string = '';
    DateOfJoining: Date = new Date();
    Salary: string = '';
}